[ahrs] int_cmpl_euler: update state interface from wrapper
